+++ R/tar_render.R
@@ -50,7 +50,7 @@
 #' @inheritParams rmarkdown::render
 #' @param path Character string, file path to the R Markdown source file.
 #'   Must have length 1.
-#' @param output Character string, file path to the rendered output file.
+#' @param output_file Character string, file path to the rendered output file.
 #' @param working_directory Optional character string,
 #'   path to the working directory
 #'   to temporarily set when running the report.
@@ -123,7 +123,7 @@
 tar_render <- function(
   name,
   path,
-  output = NULL,
+  output_file = NULL,
   working_directory = NULL,
   tidy_eval = targets::tar_option_get("tidy_eval"),
   packages = targets::tar_option_get("packages"),
@@ -142,6 +142,9 @@ tar_render <- function(
 ) {
   targets::tar_assert_package("rmarkdown")
   targets::tar_assert_file(path)
+  targets::tar_assert_chr(output_file %|||% "x")
+  targets::tar_assert_scalar(output_file %|||% "x")
+  targets::tar_assert_nzchar(output_file %|||% "x")
   if (!is.null(working_directory)) {
     targets::tar_assert_file(working_directory)
   }
@@ -155,7 +158,7 @@ tar_render <- function(
     name = targets::tar_deparse_language(substitute(name)),
     command = tar_render_command(
       path,
-      output,
+      output_file,
       working_directory,
       args,
       quiet