+++ R/tar_render_raw.R
@@ -84,7 +84,7 @@
 tar_render_raw <- function(
   name,
   path,
-  output = NULL,
+  output_file = NULL,
   working_directory = NULL,
   packages = targets::tar_option_get("packages"),
   library = targets::tar_option_get("library"),
@@ -100,9 +100,9 @@ tar_render_raw <- function(
 ) {
   targets::tar_assert_package("rmarkdown")
   targets::tar_assert_file(path)
-  targets::tar_assert_chr(output %|||% "x")
-  targets::tar_assert_scalar(output %|||% "x")
-  targets::tar_assert_nzchar(output %|||% "x")
+  targets::tar_assert_chr(output_file %|||% "x")
+  targets::tar_assert_scalar(output_file %|||% "x")
+  targets::tar_assert_nzchar(output_file %|||% "x")
   if (!is.null(working_directory)) {
     targets::tar_assert_file(working_directory)
   }
@@ -112,7 +112,7 @@ tar_render_raw <- function(
     name = name,
     command = tar_render_command(
       path,
-      output,
+      output_file,
       working_directory,
       render_arguments,
       quiet